requirements.txt
aiortc==1.6.0
bitstring==4.1.4
cryptography==41.0.7
debugpy==1.8.0
dbus-python==1.3.2
haversine==2.8.1
MAVProxy==1.8.70
numpy==2.0.0
pymavlink==2.4.41
pyOpenSSL==23.3.0
pycairo==1.25.1
PyGObject==3.46.0
pyreadline==2.1
pyrtcm==1.1.1
pyserial==3.5
pyserial-asyncio==0.6.0
pyubx2==1.2.43
PyYAML==6.0.1
requests==2.31.0
websockets==12.0

#bluezero==0.7.1
#pexpect==4.8.0
#RPi.GPIO==0.7.0 # NOTE -- I think that https://github.com/gpiozero/gpiozero may be a more popular library to be using for Raspberry Pi stuff